requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
fastapi>=0.115.5
uvicorn>=0.32.0
google-cloud-aiplatform>=1.70.0


